Factor out submit handler hook
